fix: MSVC min/max macro collision and add vcpkg binary caching
- Add NOMINMAX and WIN32_LEAN_AND_MEAN defines globally for WIN32 builds
  to prevent windows.h min/max macros from breaking std::min/std::max
- Fix unqualified min() call in QueueItem.cpp to use std::min()
- Fix FloodMap typedef in NmdcHub.h: uint32_t -> uint64_t to match
  GET_TICK() return type (eliminates narrowing warning)
- Enable vcpkg binary caching via GitHub Actions cache (x-gha) on both
  build and packaging jobs to avoid 30-minute rebuilds
- Remove debug gettext file listing step (no longer needed)

+++ dcpp/NmdcHub.h
@@ -95,7 +95,7 @@ private:
     string lastMyInfoC;
     string lastMyInfoD;
 
-    typedef list<pair<string, uint32_t> > FloodMap;
+    typedef list<pair<string, uint64_t> > FloodMap;
     typedef FloodMap::iterator FloodIter;
     FloodMap seekers;
     FloodMap flooders;
